+++ Package.swift
@@ -52,7 +52,7 @@ if let corePath = localCoreExtension {
 let package = Package(
     name: packageName,
     platforms: [
-        .iOS(.v13),
+        .iOS(.v15),
         .macOS(.v12),
         .watchOS(.v9)
     ],